Public snippets added, author row added and minor clean up
Added a public snippets page so if users want to share snippets and set
them to public they will be shown on this page.

Added author row to db, so the names of the people who created the
snippet are shown.

Did do a little hack to get the fullname, dont know if its the best way
but ill have a think.

Cleaned up various default sentry files which we dont need as of right
now.